Loaded Files
Ripped these straight from my homepage and took out some of the meat
from the index. I hope to work on this in the future and maybe make it
more generic and easy to plug in anywhere. Known bug with painting then
scrolling and painting again (coordinates are not refreshed and so brush
placement is incorrect).